requirements.txt
pandas
numpy
scikit-learn
faker
great-expectations
streamlit
matplotlib
seaborn
sqlalchemy
pytest
python-dotenv
scipy
